Fix preferences circular dependency
Fix circular dependencies present in `@theia/core/preferences` which affect the following files:
1. `preference-contribution`
2. `preference-provider`
3. `preference-service`

The issue is that all three files made use of `PreferenceScope` which was present in the `preference-service`.
The fix includes moving both the `PreferenceScope` interface and namespace to a seperate file so that the coupling
is avoided.

Circular dependencies are generally bad for a number of reasons including coupling,
prevent static linking and affect dependency injection among other reasons.
