+++ packages/core/src/browser/preferences/index.ts
@@ -18,3 +18,4 @@ export * from './preference-service';
 export * from './preference-proxy';
 export * from './preference-contribution';
 export * from './preference-provider';
+export * from './preference-scope';
