+++ packages/core/src/browser/preferences/preference-service.ts
@@ -24,54 +24,9 @@ import { Deferred } from '../../common/promise-util';
 import { PreferenceProvider, PreferenceProviderDataChange, PreferenceProviderDataChanges } from './preference-provider';
 import { PreferenceSchemaProvider, OverridePreferenceName } from './preference-contribution';
 import URI from '../../common/uri';
+import { PreferenceScope } from './preference-scope';
 
-export enum PreferenceScope {
-    Default,
-    User,
-    Workspace,
-    Folder
-}
-
-export namespace PreferenceScope {
-    export function is(scope: any): scope is PreferenceScope {
-        return typeof scope === 'number' && getScopes().findIndex(s => s === scope) >= 0;
-    }
-
-    export function getScopes(): PreferenceScope[] {
-        return Object.keys(PreferenceScope)
-            .filter(k => typeof PreferenceScope[k as any] === 'string')
-            .map(v => <PreferenceScope>Number(v));
-    }
-
-    export function getReversedScopes(): PreferenceScope[] {
-        return getScopes().reverse();
-    }
-
-    export function getScopeNames(scope?: PreferenceScope): string[] {
-        const names: string[] = [];
-        const allNames = Object.keys(PreferenceScope)
-            .filter(k => typeof PreferenceScope[k as any] === 'number');
-        if (scope) {
-            for (const name of allNames) {
-                if ((<any>PreferenceScope)[name] <= scope) {
-                    names.push(name);
-                }
-            }
-        }
-        return names;
-    }
-
-    export function fromString(strScope: string): PreferenceScope | undefined {
-        switch (strScope) {
-            case 'application':
-                return PreferenceScope.User;
-            case 'window':
-                return PreferenceScope.Workspace;
-            case 'resource':
-                return PreferenceScope.Folder;
-        }
-    }
-}
+export { PreferenceScope };
 
 export interface PreferenceChange {
     readonly preferenceName: string;
